feat: add features to hotel

+++ db/db_hotel.py
@@ -1,14 +1,25 @@
 from fastapi import HTTPException, status
+from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm.session import Session
 from sqlalchemy.testing.provision import update_db_opts
 
-from db.models import DbHotel, DbHotelManager
+from db.enums import Role
+from db.models import DbHotel, DbHotelManager, DBFeature
 from schemas import HotelBase, RateBase, UserBase
-from utils import is_unique_constraint_error, is_not_found
 
 
 def create_hotel(request: HotelBase, current_user: UserBase, db: Session):
+    if current_user.role == Role.GUEST:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
     manager = db.query(DbHotelManager).filter(DbHotelManager.user_id == current_user.id).first()
+    features = db.query(DBFeature).filter(
+        DBFeature.id.in_(request.feature_ids)
+    ).all()
+    if len(features) != len(request.feature_ids):
+        raise HTTPException(
+            status_code=400,
+            detail="One or more feature IDs are invalid"
+        )
     new_hotel = DbHotel(
         hotel_name=request.hotel_name,
         description=request.description,
@@ -17,16 +28,15 @@ def create_hotel(request: HotelBase, current_user: UserBase, db: Session):
         city =request.city,
         country =request.country,
         postcode =request.postcode,
-        manager = manager
-        #manager_id =manager.id
+        manager = manager,
+        features = features
     )
     try:
         db.add(new_hotel)
         db.commit()
         db.refresh(new_hotel)
         return new_hotel
-    except Exception as e:
-        if is_unique_constraint_error(e):
+    except IntegrityError:
             # DB rejected the transaction and raised IntegrityError, the transaction is "failed" but still open; any further queries or commits will fail until the session state is fixed
             db.rollback()
             #this clears the failed state in the session, resets the cnx
@@ -49,32 +59,35 @@ def get_all_hotels(db: Session):
     hotels = db.query(DbHotel).all()
     return hotels
 
-def update_hotel(hotel_id, request, db: Session):
+def update_hotel(hotel_id, request, current_user:UserBase, db: Session):
+    if current_user.role == Role.GUEST:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+    hotel = db.query(DbHotel).filter(DbHotel.id == hotel_id).first()
+    if not hotel:
+        raise HTTPException(status_code= status.HTTP_404_NOT_FOUND, detail="Hotel not found")
+    if current_user.role == Role.HOTEL_MANAGER and hotel.manager_id != current_user.manager.id:
+        raise HTTPException(status_code= status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
     try:
-        updated_hotel=(
-            db.query(DbHotel).filter(DbHotel.id == hotel_id).update(request.model_dump(exclude_unset=True))
-        )
-        if not updated_hotel:
-            raise HTTPException(status_code= status.HTTP_404_NOT_FOUND)
+        db.query(DbHotel).filter(DbHotel.id == hotel_id).update(request.model_dump(exclude_unset=True))
         db.commit()
         return db.query(DbHotel).filter(DbHotel.id == hotel_id).first()
-    except Exception as e:
+    except IntegrityError:
         db.rollback()
-        if is_unique_constraint_error(e):
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Hotel name and address already exists")
-        if is_not_found(e):
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Hotel not found")
-        else:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Hotel not found")
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Hotel name and address already exists")
 
-def delete_hotel(hotel_id: int, db: Session):
+def delete_hotel(hotel_id: int, current_user:UserBase, db: Session):
+    if current_user.role != Role.ADMIN:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
     hotel = db.query(DbHotel).filter(DbHotel.id == hotel_id).first()
     if not hotel:
         raise HTTPException(status_code= status.HTTP_404_NOT_FOUND, detail="Hotel not found")
     db.delete(hotel)
     db.commit()
 
-def rate_hotel(hotel_id: int, request: RateBase, db: Session):
+def rate_hotel(hotel_id: int, request: RateBase, current_user:UserBase ,db: Session):
+    if current_user.role != Role.GUEST:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+
     hotel = db.query(DbHotel).filter(DbHotel.id == hotel_id).first()
     if not hotel:
         raise HTTPException(status_code= status.HTTP_404_NOT_FOUND, detail="Hotel not found")
@@ -93,4 +106,6 @@ def get_rating(hotel_id: int, db: Session):
     hotel= db.query(DbHotel).filter(DbHotel.id == hotel_id).first()
     if not hotel:
         raise HTTPException(status_code= status.HTTP_404_NOT_FOUND, detail="Hotel not found")
+    if hotel.rating_count == 0:
+        return 0
     return round(hotel.rating_sum/hotel.rating_count,1)
